[Afrath]Updated the file with test case to check double

+++ GenericsAndTesting/FindThreeNumbers.cs
@@ -8,7 +8,7 @@ namespace GenericsAndTesting
 {
     public class FindThreeNumbers
     {
-        public int FindMaxNumber(int number1, int number2, int number3)
+        public double FindMaxNumber(double number1, double number2, double number3)
         {
             if ((number1.CompareTo(number2) > 0) && (number1.CompareTo(number3) > 0))
             {
